backend/requirements.txt
# FastAPI and Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic and Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.0.0

# Authentication
fastapi-users[sqlalchemy]==12.1.2
fastapi-users-db-sqlalchemy==6.0.1

# File Storage
boto3==1.29.7
minio==7.2.0

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# Celery (Phase 2)
celery==5.3.4

# Utilities
pytz==2023.3
python-dateutil==2.8.2
